move lots of deserialization code into new module
